Create Euler Problem 2- Sum of even Fibonacci numbers.py
Uploading my solution for Euler Project Problem #2. My solution uses an existing generative function that I developed to yield a set number of items in the Fibonacci sequence. I modified that function to also take into account an upper bound (value limit), which in this case is set to 4,000,000 per this specific problem.

As usual, there is likely a more concise way to do this. But, this implementation is very readable.