Euler Problem 7- 10,001st Prime
This initial commit contains unused lines of code for my own personal record. At one point, I thought I might increase efficiency by storing and checking primes previously found against the number actively being checked. In the end that was not faster. Below are the times I recorded when testing variations:

101st prime
-----------
With the list of primes...
This program took 5.7712414264678955 seconds to run.

Without the list of primes...
This program took 3.918135643005371 seconds to run.

Without the print statements...
This program took 0.02094578742980957 seconds to run.

1001st prime
------------
7927 is the 1001st prime.
This program took 0.26682615280151367 seconds to run.

10,001st prime
--------------
104743 is the 10,001st prime.
This program took 24.218202590942383 seconds to run.

Note: This commit also contains a minor formatting change for Euler Problem 3, to make it consistent with my newer solutions. I happened to have it open when I was considering how to check for primes.