Create Euler Problem 8- Largest Product in a Series.py
This initial commit contains takes in the original, arbitrary 1,000-digit number and converts it into a string to make it iterable. Then it builds an n-digit list from the first n digits, each being converted to an int so they can be multiplied. Each n-digit list is multiplied and compared to the last, with the greatest product and its factors saved and then printed at the end. There are several print statements in the comments that can be activated to demonstrate how the algorithm works; however, this slows it down. As is, the program runs in about .07 seconds on my machine, when running the original problem as specified.