Create Euler Problem 9- Special Pythagorean triple.py
This solution is adapted from a an excellent and well-explained solution found at:
https://blog.dreamshire.com/solutions/project_euler/project-euler-problem-009-solution/
Their answer remains intact. I have adjusted for readability upon output and proof that the numbers fit Pythagoras' theorem. My comments in the code file are there to summarize the code for learners, such as myself.

Note: My original solution used a formula that can generate an infinite number of pythagorean triples if you feed it any rational positive number. An explanation can be found in this video:
https://www.youtube.com/watch?v=y718ckf336c

3Blue1Brown has a beautiful video visualizing all possible Pythagorean triples. It can be found at:
https://www.youtube.com/watch?v=QJYmyhnaaek