Create Euler Problem 10- Summation of Primes.py
This solution is based on the Sieve of Erastosthenes. You can find an explanation of this method and the pseudocode this algorithm is based upon at:
https://en.wikipedia.org/wiki/Sieve_of_Eratosthenes

This code creates two lists: one called "numbers" that is a mere list of all integers below a user-specified number, n, and another called "sieve", which contains booleans, set to True. Using the Sieve of Erasthones method, all booleans in "sieve" that lie at composite-numbered list indexes are changed to False. Then, each number in "numbers" is printed if that number corresponds to True in the sieve list.

Benchmarks:
*At the time of uploading to GitHub, my machine computes the standard test case of 2,000,000 in less than half a second.
*7,000,000 solves in 1.86 seconds.
*20,000,000 solves in 5.42 seconds.