Create Euler Problem 12- Highly Divisible Triangular Number.py
This solution uses a generator to produce triangular numbers, starting at one and going until the user-entered amount of divisors is exceeded. A fairly easy-to-read explanation of generators can be found at:
https://realpython.com/introduction-to-python-generators/
("Example 2: Generating an Infinite Sequence" was useful in this instance.

My thanks to reddit user agf and those who edited agf's explanation and code showing a concise way to find all factors of a given number. That can be found at:
https://stackoverflow.com/questions/6800193/what-is-the-most-efficient-way-of-finding-all-the-factors-of-a-number-in-python/

Benchmarks:
*At the time of uploading to GitHub, my machine computes the standard test case of 500 divisors about 2.96 seconds.
*750 divisors was solved in 9.02 seconds.
*1,000 divisors was solved in 32.07 seconds.