Create Euler Problem 13- Large Sum.py
This solution is actually fairly easy to understand, despite the size of the numbers in the problem. The code could be leaner, but it it is highly readable as is.

This solution takes in the original list of 50-digit numbers as one long string delimited by new lines. The data is cleaned up by splitting the string into a list of 50-digit numbers (keeping them formatted as strings), and then converting each number into a sub-list made up of five 10-digit numbers (as strings) so that we can sum the numbers in 10-digit chunks. This gets around the problem of dealing with really long numbers.

The first ten digits of each number are added and saved in a list_of_totals. Then, the same is done for the second ten digits and so on.

Then, just like in grade-school addition, the leftmost digits (in this case starting with the 10e9th place value) are "carried" to the left and added to the next 10-digit chunk.

I have no doubt there are more efficient solutions (at least in terms of having fewer lines of code). However, the approach is easily explainable, and the code is highly readable.