Create Euler Problem 14- Longest Collatz sequence.py
When reading Problem 14 on Project Euler, I realized something -- I had already solved this. So, this solution is adapted from prior code I posted on GitHub in November 2022.

A summary of Collatz conjecture:
In 1937, a German mathematician named Lothar Collatz formulated a conjecture (it still remains unproven) that can be described in the following way:
    1. Take any non-negative and non-zero integer number greater than 1 and name it c0.
    2. If it's even, evaluate a new c0 as c0 ÷ 2.
    3. Otherwise, if it's odd, evaluate a new c0 as 3 × c0 + 1.

The hypothesis/conjecture says that regardless of the initial value of c0, it will always go to 1.

The program below takes one natural number and executes the above steps as long as c0 remains different from 1. It also counts the steps needed to achieve the goal. The program continues to run until it has checked all sequences up to the number specified in the code (variable: max_attempts). The output prints the maximum number of steps in a Collatz sequence and the number that produced that sequence.

If you would like to see all the sequences produced, you can un-comment several print statements within the code.

Benchmarks:
*At the time of uploading to GitHub, my machine tests all cases up to 1,000,000 in approximately 21.38 seconds.
*3,000,000 cases were tested in approximately 69.26 seconds.