Create Euler Problem 15- Lattice Paths
This solution treats a grid like a portion of a Pascal's triangle (but rotated to 45 degrees).

First, it takes a user-specified number, n, representing the size of a grid. It then makes a grid of numbers that looks like this:

0 1 1 1 1 ...
1 0 0 0 0 ...
1 0 0 0 0 ...
1 0 0 0 0 ...
1 0 0 0 0 ...
...

The length of the rows and columns are n+1. The numbers on the top row and leftmost column represent the number of paths from the top-left origin to that point on the grid. The rest of the zeroes are placeholders.

Next, treating the grid like a Pascal's triangle, each of the interior zeroes is replaced with the sum of the digit above and the digit to the left. This proceeds from left to right, top to bottom. The number on the bottom right of the grid is the total number of lattice paths that can be made to that point.

Benchmarks:
*Despite the size of the code, it takes less than 0.01 seconds to execute either the Euler problem case of a 20x20 grid or a test case of 100x100.