Create Euler Problem 16- Power Digit Sum
This solution is straightforward and can be read in the code. The user provides a base number (x) and exponent (n). Then, the program calculates and prints the sum of the digits of the resulting number.

Some lines of the code can be un-commented, resulting in the printing of intermediate steps (such as the product of x to the nth and the intermediate sum of all digits checked at each step).

Benchmarks:
*As of the date of upload to GitHub, this code takes less than 0.01 seconds to execute any test case with an x and n below 1,001.