Create Euler Problem 17- Number Letter Counts.py
This solution leverages the num2words module. (More info at: https://pypi.org/project/num2words/)

The user provides a maximum number, n. For each integer from 1 to that number (inclusive), num2words converts that integer into a string and concatenates it to a string_of_nums. Then, all spaces and hyphens are removed, leaving us with a string of letters, the length of which is printed.

Benchmarks:
*At the time of uploading to GitHub, my machine computes the standard test case of all numbers 1 to 1,000 in about 0.099 seconds.
*1 to 10,000 is solved in 0.49 seconds.
*1 to 100,000 is solved in 6.06 seconds.