Create Euler Problem 20- Factorial Digit Sum.py
Python makes this problem easy, as it can handle numbers with as many digits as you need in order to perform a calculation. It also has a factorial function in the math module. This solution sums the digits of the factorial of an arbitrary integer in four lines of code. But, you can do it in fewer if you really need to.