Create Euler Problem 67- Maximum Path Sum 2.py
A follow up to Euler Problem #18, this solution uses the same code on a list with 100 row and still solves it in a fraction of a second. For more information, see annotations in the code as well as the code for Euler Problem 18 at:
https://github.com/soloudithz/Project-Euler-solutions

---
Housekeeping Note: I renamed the text file from Problem 18 to be consistent with the naming convention of the text file for Problem 67. So, this commit also includes that file.