Create Euler Problem 21- Amicable Numbers.py
This is my first attempt at Problem 21. The solution checks each number from two to a user-specified number (inclusive) to find the sum of that number's factors. Then, the sum of that derived sum's factors is checked to see if it is equal to the first number. If so, both numbers are appended as a tuple to a list, amicable_nums. Once all amicable pairs are found, each pair is summed and their combined total is printed.

Benchmarks:
--------------
At the date of this commit, the program computes the Project Euler test of all numbers up to 10,000 in approximately 16 seconds. I am working to get the time down.