Create Euler Problem 22- Name Scores.py
Reference: https://projecteuler.net/problem=22

This solution reads in the pe022_names.txt file and cleans up the data a bit. Then it creates a dictionary of values for letters of the alphabet. Each name is scored based on the sum of the values of its letters, and that score is append to a list of name_scores. Finally, it prints the result.

Benchmarking:
-----------------
Some programmers prefer readability and others prefer conciseness, and then others value efficiency at runtime (i.e., speed of processing). I have always favored readability unless it significantly reduces performance in a case where performance is a deal-breaker. So, for benchmarking purposes, this commit includes the following files:

* pe022_names.txt (the Project Euler text file needed to solve the problem
* my solution code file ("Euler Problem 22- Name Scores.py")
* an excellent and concise but less human-readable alternative solution ("Euler Problem 22- Name Scores (dreamshireversion.py")
* two screenshots documenting the speed of the code at runtime

In this case, my code, which was written to be more human-readable, was observed to also run more about eight times more quickly than the alternative (concise) solution. Feel free to test this on your own machine.